+++ src/gdal_utils.hpp
@@ -33,6 +33,8 @@ GDAL_ASYNCABLE_GLOBAL(info);
 GDAL_ASYNCABLE_GLOBAL(translate);
 GDAL_ASYNCABLE_GLOBAL(vectorTranslate);
 GDAL_ASYNCABLE_GLOBAL(warp);
+GDAL_ASYNCABLE_GLOBAL(buildvrt);
+GDAL_ASYNCABLE_GLOBAL(rasterize);
 
 } // namespace Utils
 } // namespace node_gdal